Refactor: separate files with packages
기능 별로 패키지를 만들어, 파일을 분리하였습니다.
