+++ src/main/java/ladder/creator/LadderCreator.java
@@ -1,4 +1,7 @@
-package ladder;
+package ladder.creator;
+
+import ladder.Ladder;
+import ladder.wrapper.NaturalNumber;
 
 public interface LadderCreator {
     Ladder createLadder(NaturalNumber numberOfRows, NaturalNumber numberOfPerson);
