+++ src/main/java/ladder/creator/RandomLadderCreator.java
@@ -1,4 +1,8 @@
-package ladder;
+package ladder.creator;
+
+import ladder.Ladder;
+import ladder.wrapper.NaturalNumber;
+import ladder.wrapper.Position;
 
 import java.util.Random;
