+++ src/main/java/ladder/game/LadderGameFactory.java
@@ -1,4 +1,8 @@
-package ladder;
+package ladder.game;
+
+import ladder.wrapper.NaturalNumber;
+import ladder.creator.LadderCreator;
+import ladder.creator.RandomLadderCreator;
 
 public class LadderGameFactory {
     public static LadderGame createRandomLadderGame(NaturalNumber numberOfRows, NaturalNumber numberOfPerson) {
