+++ src/main/java/ladder/printer/Printer.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.printer;
 
 public interface Printer {
     String print();
