+++ src/main/java/ladder/util/Direction.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.util;
 
 public enum Direction {
     LEFT(-1), NONE(0), RIGHT(1);
