+++ src/test/java/ladder/RowTest.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.wrapper.NaturalNumber;
+import ladder.wrapper.Position;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
